updated sql script - reimb was being added when no such employee in users table

+++ src/main/java/com/revature/service/Validator.java
@@ -67,7 +67,7 @@ final class Validator {
         else throw new MyException("Invalid reimbursement type provided");
     }
 
-    public static boolean isValidAmount(int amount) {
+    public static boolean isValidAmount(double amount) {
         if(amount > 0) {
             return true;
         }
